Ongoing: move unit tests to main source tree and run them via an embedded SpecRunner.
Reasoning: in dev mode, the database is re-created on each start of run-app. So we have a stable test environment. The advantage of running the unit tests vs a real database is that development is easier and if it runs this way as an "integrated unit test", I can expect the production code to run as well. I tried to use mocking (@TestFor, @Mock), and at first and it was very helpful and easy. Until I encountered strange Hibernate persistence issues with mocked objects which were impossible to debug within reasonable time.